audio_extn: fix mute for HFP back to back call
Internal variable which holds the mute state
for HFP client call was not reset after the
call ended and as a result the state was carry
forwarded to subsequent call. If first call
was muted, then mute for subsequent call was
ignored, not forwarded to PAL in audio_extn
(as the state is already mute) but Dialer app
shows muted on UI.

Fix is to reset the internal state before starting
the HFP client call.

Change-Id: I99300ab9a0ebf94996e6635d35d5eb93e804cfe6