core: handle read/standby properly for non-lab hotword record
Sometimes hotword record may be used for recording usecase like
okg enrollment, in this case read and standby is not working
as expected because hotword record was expected to be for VA
lab recording only.

Add flag to indicate if hotword record is for lab record or not,
so that we can handle read/standby correspondingly.

Change-Id: Ic95f81b47e5d26dd93d93b0fc6790caf1f7f4dce
(cherry picked from commit d7a7a30ef27643ec01c86be30bbdafe3b947a8c7)