core: handle DRAIN_ALL command in DRAINING state
Even though, this command is not allowed as per async
stream state machine. But this is potential command
which HAL needs to handle.

This is missed in Android V interface agreement.
Anyway, we don't have the VTS tests for compress
offload playback, so it's safe to maintain.

This resolves issue seek bar progressing beyond the length
of the clip.

Change-Id: I937d3941ebf707e38314ac5bf271741d33925c29
(cherry picked from commit 383daa95ad2864b16481d7f973186a1ad3cab9b1)