core: set None Device instead of empty device for stream
This change is same df597fc3c6c46461aa03082c4144359da603412f
commit.
Here is the summary.

None Device Fail-Safe

Although setting empty devices on stream is allowed momentarily.
But read's or write's to stream when empty devices lead to failures.

Configure stream devices to the NONE device as a fail-safe mechanism.
This ensures that any attempts to read from or write to the stream when no device is
connected are handled gracefully.

Note: This scenario is not expected to occur. The HAL client (i.e., framework) must
ensure that this situation does not arise.

PAL layer can also adjust its behavior, based on this.

Change-Id: I7292cf778e164ee033798137b8ea9de691a72c87