core: Fix bitwidth for pcm offload
FWK generally sends bit_width from PCM encoding, but the data format written to HAL can differ from the encoding. While configuring PAL, offload_info's bit width is used for offload cases; however, it might not always be in sync.

For example, when playing a 24-bit WAV file, the encoding is set to 24 by the extractor, but the AOSP extractor provides float output, so it is written as float data.

To address this:
- If the format is PCM, use the bit width from the format.
- For compressed formats, rely on the bit width from offload_info.
cherrypick of I0ebbf80973dcdbecd25dff586bc23197cd535b5b

Change-Id: I80aa8a6683e528872a8b93c553ae51c0821ff67f