effects: Fix preset enablement
Reverb PRESET_NONE is equivalent to disabled state,
But support for this state is not provided in DSP.
Hence, do not set enable flag, if in peset mode with preset NONE.
Effect would be enabled when valid preset is set.
Also, add proper range checks to avoid out of bound access.

Change-Id: I364570073b76ebab0f0207ea0220ae193513e386