hal: change defaults for fd tracker
Change default abort value for leak tracker, 1024
open fd count is more than sufficient to detect / assume a leak.

Enable on debug builds by default to root cause
the issues properly, although it can have some performance overhead if the leak thread is running by default on debug builds, but as it runs only once in 180seconds (which can be configured as well), so impact should be negligible.

Change-Id: I8f97d40c8b778ea4c645436cbc7be77a9d188329