hal: Improve bootup logging
"vendor_audio_interfaces.xml" is key config
file to load and register various services.
In absence of this file, HAL is not expected
to behave in proper manner, if the file doesn't
exist crash ASAP to validate the configuration.
This is to catch setup issues on new targets.

IModule/default should be registered from libaudiocorehal.qti
to cater internal HAL. However, if it is not registered
from mentioned libs then it is tried as fallback from
libaudiocorehal.default, as it is necessary interface for
framework bootup. However, it shouldn't come from both config files
ie. audio_module_config_primary.xml and audio_policy_configuration.xml.
For initial setup of targets, it can be seen, so better crash
if multiple XML are trying to register this interface.

In case when HAL monitor expires, dump the status of
needful services to check debug points.

Change-Id: I2884531db9a58bf69aceaeb2961680127cd0c335